etl/requirements.txt
requests>=2.31.0
snowflake-connector-python>=3.6.0
python-dotenv>=1.0.0
